+++ src/bot/enums/AheadEnum.ts
@@ -1,0 +1,6 @@
+enum AheadEnum {
+  INSIDE = 'INSIDE',
+  OFFSIDE = 'OFFSIDE',
+}
+
+export default AheadEnum;